Task 10

+++ 0x05-python-exceptions/103-python.c
@@ -106,8 +106,8 @@ void print_python_float(PyObject *p)
 
     value = float_object->ob_fval;
 
-    if (value == (int long)value)
-        printf("  value: %.1f\n", float_object->ob_fval);
+    if (value == (int long long)value)
+        printf("  value: %#.1g\n", float_object->ob_fval);
     else
         printf("  value: %.16g\n", float_object->ob_fval);
